fixed recursive deps and refactored and insert only one app into overview at a time
